AddtionalInformation, EmailValidation 컴포넌트의 분리
SignUp컴포넌트에 과도하게 치중된 책임을 AddtionalInformation, EmailValidation컴포넌트로 분리함으로 써 해결.
